fix(db): ensure anomaly_thresholds table is created by Flyway migration

+++ backend/src/main/java/vn/nguongocso/alert/dto/request/CategoryThresholdOverrideRequest.java
@@ -1,5 +1,6 @@
 package vn.nguongocso.alert.dto.request;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import jakarta.validation.constraints.DecimalMin;
@@ -34,7 +35,7 @@ public class CategoryThresholdOverrideRequest {
 
     @NotNull(message = "Khoảng cách tối đa cho phép không được để trống")
     @DecimalMin(value = "0.0", message = "Khoảng cách tối đa cho phép phải không âm")
-    private Double maxDistanceKmPer30Min;
+    private BigDecimal maxDistanceKmPer30Min;
 
     @NotNull(message = "Thời gian tối thiểu giữa các lượt quét không được để trống")
     @Min(value = 0, message = "Thời gian tối thiểu giữa các lượt quét phải không âm")
